feat: MapItem 분리, 좋아요기능
